Scripts: Remove createRandom, its done in scripts now Cleanup warnings

+++ Source/Fodder.hpp
@@ -1277,8 +1277,8 @@ public:
 
     void            String_Print_Small(std::string pText, const size_t pY);
     void            String_Print_Small(std::string pText, const size_t pX, const size_t pY);
-    void            String_Print_Large(std::string pText, const bool pOverAndUnderLine, const uint16 pY);
-	void            String_Print_Large(std::string pText, const bool pOverAndUnderLine, const uint16 pX, const uint16 pY);
+    void            String_Print_Large(std::string pText, const bool pOverAndUnderLine, const size_t pY);
+	void            String_Print_Large(std::string pText, const bool pOverAndUnderLine, const size_t pX, const size_t pY);
 
 	void			DataNotFound();
     virtual void    Prepare(std::shared_ptr<sFodderParameters> pParams);
